requirements.txt
scapy-python3
matplotlib
numpy
scipy

